// src/ast-types.ts
export namespace ExprNS {
    export interface Visitor<T> {
        visitBigIntLiteralExpr(expr: BigIntLiteral): T
        visitBinaryExpr(expr: Binary): T
        visitCompareExpr(expr: Compare): T
        visitBoolOpExpr(expr: BoolOp): T
        visitGroupingExpr(expr: Grouping): T
        visitLiteralExpr(expr: Literal): T
        visitUnaryExpr(expr: Unary): T
        visitTernaryExpr(expr: Ternary): T
        visitLambdaExpr(expr: Lambda): T
        visitMultiLambdaExpr(expr: MultiLambda): T
        visitVariableExpr(expr: Variable): T
        visitCallExpr(expr: Call): T
        visitComplexExpr(expr: Complex): T
        visitNoneExpr(expr: None): T
    }
    export abstract class Expr {
        startToken: Token;
        endToken: Token;
        protected constructor(startToken: Token, endToken: Token) {
            this.startToken = startToken;
            this.endToken = endToken;
        }
        abstract accept(visitor: Visitor<any>): any;
    }
// ...
    export class Variable extends Expr {
        name: Token;
        constructor(startToken: Token, endToken: Token, name: Token){
            super(startToken, endToken)
            this.name = name;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitVariableExpr(this)
        }
    }
// ...
}
export class ImportAlias {
    constructor(public name: Token, public alias: Token | null) {}
}

export namespace StmtNS {
// ...
    export abstract class Stmt {
        startToken: Token;
        endToken: Token;
        protected constructor(startToken: Token, endToken: Token) {
            this.startToken = startToken;
            this.endToken = endToken;
        }
        abstract accept(visitor: Visitor<any>): any;
    }
// ...
    export class Assign extends Stmt {
        name: Token;
        value: ExprNS.Expr;
        constructor(startToken: Token, endToken: Token, name: Token, value: ExprNS.Expr){
            super(startToken, endToken)
            this.name = name;
            this.value = value;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitAssignStmt(this)
        }
    }
// ...
}